A psychology survey as a csv prjoect.
A Psychology survey conduct by me and my friends, and presented in a csv file for class 12th project, with additional functions as to append and delete data; would be a very unique idea if you want to present it as a project to your school; something more than those hotel management programs,  or you could just  have fun. It depends on you. 

KEEP IN MIND TO CHANGE THE NAME ! THE INVIGILATOR COULD BE FORGIVING AND NAIVE, BUT HE/SHE IS NOT STUPID.
CONTACT IF YOU MAY WISH.

GIT-CLONE IT AND ENJOY!